+++ packages/assessment-symbol-search/src/runner.ts
@@ -6,8 +6,18 @@
 import { Session } from "@m2c2kit/session";
 import { SymbolSearch } from "./index";
 
+const activity = new SymbolSearch();
+
+// Set game parameters from URL parameters
+const urlParams = new URLSearchParams(window.location.search);
+const gameParameters: { [key: string]: string } = {};
+urlParams.forEach((value, key) => {
+  gameParameters[key] = value;
+});
+activity.setParameters(gameParameters);
+
 const session = new Session({
-  activities: [new SymbolSearch()],
+  activities: [activity],
 });
 
 session.onActivityData((ev) => {