+++ packages/schematics/src/new/index.ts
@@ -193,7 +193,7 @@ function generateAppPackageJson(name: string) {
   "dependencies": {
     "@m2c2kit/core": "${Constants.M2C2KIT_CORE_PACKAGE_VERSION}",
     "@m2c2kit/addons": "${Constants.M2C2KIT_ADDONS_PACKAGE_VERSION}",
-    "@m2c2kit/session": "^${Constants.M2C2KIT_SESSION_PACKAGE_VERSION}"
+    "@m2c2kit/session": "${Constants.M2C2KIT_SESSION_PACKAGE_VERSION}"
   },
   "devDependencies": {
     "@m2c2kit/build-helpers": "${Constants.M2C2KIT_BUILD_HELPERS_PACKAGE_VERSION}",
@@ -242,10 +242,10 @@ function generateModulePackageJson(name: string) {
   ],
   "dependencies": {
     "@m2c2kit/core": "${Constants.M2C2KIT_CORE_PACKAGE_VERSION}",
-    "@m2c2kit/addons": "${Constants.M2C2KIT_ADDONS_PACKAGE_VERSION}",
-    "@m2c2kit/session": "^${Constants.M2C2KIT_SESSION_PACKAGE_VERSION}"
+    "@m2c2kit/addons": "${Constants.M2C2KIT_ADDONS_PACKAGE_VERSION}"
   },
   "devDependencies": {
+    "@m2c2kit/session": "${Constants.M2C2KIT_SESSION_PACKAGE_VERSION}",
     "@m2c2kit/build-helpers": "${Constants.M2C2KIT_BUILD_HELPERS_PACKAGE_VERSION}",
     "@m2c2kit/schema-util": "${Constants.M2C2KIT_SCHEMA_UTIL_PACKAGE_VERSION}",
     "@rollup/plugin-node-resolve": "${Constants.ROLLUP_PLUGIN_NODE_RESOLVE_VERSION}",